BM-569: Add retries for batch submission (#340)
Previously if a batch submission transaction failed for any reason (e.g.
a connection reset on the RPC) the broker would never attempt to
resubmit the batch and all tasks in that batch would fail.

This adds some resilience by adding a configurable parameter for number
of attempts to resubmit a batch. Batches that fail will retry for the
given number of attempts before setting their status to Failed.

- Adds new config field for BatcherConfig - max_submission_attempts
(default=3)
- Adds test for multiple retries timing out

> Note: This solution avoids storing the number of submission attempts
in the DB and so is not resilient to failures that trigger a panic and
restart. In this case the batch will still be dropped. If this comes up
as in issue in the future we can move the storage of submission attempts
to the db