BM-626: Fix bento crash with zero keccak input request (#396)
The error would panic and crash the instance, which was indefinitely
irrecoverable. This rejects that invalid case, such that the node will
continue.

The part I am unsure about is the sanity check within the executor for
keccak requests. This changes to error log and skip those requests, and
didn't seem like a clean mechanism to fail to proof as a whole, but
perhaps this should still go through to fail with the prove agent so
that it errors and then perhaps the tasks get cancelled through taskdb?
Haven't looked that in depth at the implications yet.