BM-657: Add retries for calls to Bonsai (#415)
We've seen transient errors when calling the Bonsai API. This adds
retries to all API calls (previously we only retried calls to check job
status). This also adds retries to the entire proving workflow, to
better capture the case where are able to kick off a proving job and it
terminates, but it terminates with a retryable internal server error.

This should help with:
https://github.com/boundless-xyz/boundless/issues/410

---------

Co-authored-by: Willem Olding <willemolding@gmail.com>
Co-authored-by: Victor Snyder-Graf <victor@risczero.com>
Co-authored-by: Rami Lukata <32602478+rlukata@users.noreply.github.com>
Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>