BM-598: Limit concurrent locked orders (#413)
Solves part of BM-598 in that we need multiple ways to limit the
capacity of orders taken on by the broker, and limiting the amount of
current locks is not a perfect solution but a reasonable mechanism that
would be useful in the short term or to limit risk.

Will open PR for review once more thoroughly tested and will try to make
sure no edge cases and minimimal overhead. Basically how this works is
querying the database for orders that have been selected to be locked
but not fulfilled, and then spawning tasks to handle pricing each. The
capacity is calculated on an interval, every 10s, where the config and
db will be checked, which will pick up when orders are fulfilled and
more capacity available. Whenever pricing successful, it will reduce
from the capacity to avoid extra tasks from being marked as locked.

May also refactor the design to do this, as this doesn't handle the case
where the order is planned to be locked in the future very well. The
issue with having the concurrent locks be monitored in the order monitor
(which actually calls lock) is that it isn't clear when the broker needs
to lock the order in by to make sure that it will complete it in time,
because that is done in pricing.

---------

Co-authored-by: Victor Snyder-Graf <victor@risczero.com>