BM-1210: Handle race condition between indexer and monitor (#811)
Currently a race condition between indexer and monitor causes variosu
events to be missed, which cause alarms to go off:

1. Indexer sets "now" timestamp
2. Indexer starts processing events for "now"
3. Monitor sets "now" timestamp reads from DB and publishes events for
"now"
4. Indexer writes events to DB

Note this is a quick fix. Ideally the monitor would look at the last
processed block table, and use that to avoid race conditions, but this
is a larger refactor.