BM-1307: Support fetching of ERC-1271 authorized requests in `boundless` cli (#890)
The `Order` struct used throughout the code base makes a bad assumption
that all signatures are ECDSA, and can be parsed by `alloy`. Signatures
should not be parsed, and instead should be handled as bytes expect
where they are verified. Fixing this issue is a much deeper refactor,
but this PR starts by refactoring the method to fetch an order from the
market, with the immediate goal of supporting handling of ERC1271
authorized requests in the `boundless` CLI.