Update README.md
Hello,

I see two typos here and request correction.

First one: In the sentence "It is used to store somewhat sensative API keys for RPC services and Etherscan" the word "sensative" is misspelled. It should be "sensitive", and

Second one: In the note "This contract should only be deployed if necessary as, on most cases, should be reused" the preposition "on" is used incorrectly. It should be "in".

Thx