BM-1487: [DOCS] Fix Dead Steel links (#1022)
After merging #965, I noticed this CI error:
https://github.com/boundless-xyz/boundless/actions/runs/17097732120/job/48486049494

It's because the CI on the merged PR ran for the last time before Steel
was moved into its own repo. I updated the Steel doc links to point to
the right place now so that this CI error is fixed, and docs changed can
continue to be deployed correctly.