Cap rewards based on the log ID to prevent avoidance of the cap (#1038)
Currently on `victor/povw` the reward cap is determined based on the
value recipient. This allows recipients to exceed the available reward
cap by splitting their work across multiple logs, and then producing
multiple mint transactions. This circumvents the intention of the
completeness checks to ensure that the cap is applied to the total
rewards for the epoch.

To address this issue, this PR changes the mint calculator to apply the
reward cap based on the work log ID. This aligns with the completeness
checks to ensure the cap cannot be circumvented.

As an effect of this change, provers will need to delegate the reward
power of their staked ZKC to the work log ID.