Prevent the zero address from being used as value recipient (#1055)
On `victor/povw`, there is an issue that is a prover specifies a value
recipient of zero, they will not only burn the rewards for that update,
but will be prevented from claiming rewards for any future update.

This PR fixes this by preventing a value recipient of zero from being
specified. The check is added to the log builder guest, such that the
update cannot be proven or posted to the PoVW Accounting contract.