BM-1637: [V-BND-VUL-001] Add gas check to ensure sufficient gas is available for an optional callback (#1159)
This PR introduces a gas check to ensure sufficient gas is available for
an optional callback executed when a prover fulfills a request. The
check verifies that the remaining gas, adjusted for EIP-150's 63/64
rule, meets or exceeds the `callbackGasLimit` specified by the
requestor. This prevents a prover from fulfilling a request with
insufficient gas for the callback.