BM-1757: check minio for completed receipt if job not found in taskdb (#1250)
There is currently a race condition with #1218 such that if the cleanup
happens after a job is completed but before the broker polls again, it
will return a failed status. This change checks minio for the receipt if
not found in taskdb to avoid this issue and allow clearing pgsql more
liberally in other ways also.