BM-1765: mine by default, opt out with env (#1255)
Passby also reduces the default mining job size to ~1h on 5090, I've
found that when consistently participating in the market, 4h is too long
where it would be inconsistent where it is included in the epoch. This
is an updated guess of what the default could/should be.

- Updates `just broker` to `just prover` (broker was already overloaded
given it started bento also)
- Removes `just mine` command and rolls it into `just prover`
  - `BOUNDLESS_MINING=false` to disable the mining process
- Adjusts default mining job size as mentioned above