BM-1944: Catch tx conf errors in OGs, lower sev2 thresholds for expired requests, fix error logging in monitor (#1332)
Current a single tx confirmation error will trigger an alert. Catching
and raising threshold.

Market capacity is somewhat flakey at the moment, which triggers lots of
SEV2s. Lowering it to 50% for now, since there are known issues and we
are monitoring manually.

Currently error logs are swallowed in the lambda monitor. One line
change that should include full logs.