BM-1950: Update log lambda to query alarm tags for debugging via chatbot (#1338)
This PR adds an additional approach for the log lambda to find the
information it needs either to output debug data via chatbot or to
create a log query. If an incoming alarm has the relevant information in
tags then the lambda won't need to rely on parsing the alarm names,
offering an option for debug configuration that can be more robust and
easier to expand on in the future.

In order to achieve this, the lambda needs to assume a role in the
target account that allows listing tags on the alarm resources. This is
set up here via new roles in `bootstrap` and a change to the lambda
execution role, plus new lambda code.

The original name parsing logic remains and is still used in all cases
in which alarms are not tagged (currently all cases except for
prover-cluster alarms, which add tags in this PR as well).