BM-2023: Indexer API for Explorer (#1381)
Large refactor of our indexer + exposing it as an API for the explorer.
Many breaking changes, so we will wipe the databases in Prod when this
is deployed.

Changes:
1/ Now fetches all data (events, block metadata, etc) in advance, with
various parallelization and caches to file storage / S3
2/ Support for different RPC providers for different calls. 
3/ Support for different APIs for fetching TX metadata. Either getting
the full block and parsing receipts, or fetching the transactions
individually. Different strategies are more efficient on different RPC
providers
4/ Precomputes aggregations hourly/daily/weekly/monthly
5/ Support for running backfills

Core logic to review:
crates/indexer/src/market/core.rs