BM-2095: Fix URL parsing V2. Fix distributor in staging (#1408)
There is some sort of issue with clap + the usage of custom parsers with
Option<Url> type. We were seeing error `Mismatch between definition and
access of `rpc_url`. Since we need to support empty string, due to the
way the docker compose works, and empty string is not well supported by
the Url type, the simplest option is just to use String type.

Also fixes the distributor in staging.