BM-2103: Restrict indexer monitor from querying events from start of time. Fix queries. (#1415)
Currently the monitor isn't running as its attempting to query for every
metric since the beginning of time. I think this happened when we wiped
the database and we lost track of what metrics we'd processed.

Additionally some of the queries are using a field that doesn't exist in
the fulfillment events table. Fixed this.

I think these two issues are why various alarms, e.g. Signal, are stuck
in alarm state.