BM-2148: Improved logging around proof generation and submission (#1442)
Currently its hard to tie proving errors to the orders that caused them.
This makes it super clear when proofs are run for each order id, and
when they fail/retry

This also should help debug an issue with fulfillment where we see
"RequestIsNotPricedOrLocked" reverts. Its unclear which contract
function we are trying to execute when we see this, which makes
debugging difficult.