BM-2191: Fix false positives on monitor, fix api caching (#1471)
Monitor is having false positives as the indexer is running behind by
over a minute. This means when the monitor looks for entries in the
indexer table for the past minute, the entries haven't been inserted yet
as the indexer is running behind. Increasing the monitor lag from 1 min
to 5 mins to accomodate.

Lowering HOURLY_AGGREGATION_RECOMPUTE_HOURS from 12 => 6, which should
speed up individual indexer runs.

We are seeing strange behavior on the API, specifically random looking
404s, with no associated APIG/Lambda logs. It seems like Cloudfront is
the culprit. Fixing Cloudfront caching behavior which could be causing
this.

Also assorted logging improvements/fixes